requirements.txt
certifi==2022.12.07
charset-normalizer==2.0.12
feedparser==6.0.8
html2text==2020.1.16
idna==3.3
mccabe==0.6.1
numpy==1.26.4
opensearch-py==2.4.2
pyflakes==2.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.0.3
requests>=2.27.1
sgmllib3k==1.0.0
six==1.16.0
sodapy>=2.2.0
tqdm==4.63.0
url-normalize==1.4.3
urllib3>=1.26.1
w3lib==1.22.0
